+++ CHANGELOG.md
@@ -7,6 +7,8 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 
 **New fixes in 1.27.x:**
 
+* [mse121682: Links to election nominations don't work after nominations close](http://meta.stackexchange.com/q/121682)
+
 1.26 (10 Dec 2014)
 ====